Move NotificationAction interface to proper module
