+++ src/main/typescript/notifications/actions/DismissAction.ts
@@ -2,7 +2,7 @@ import { GitHub } from '@github-manager/client';
 import { Logger } from '@github-manager/utils';
 import { _ } from '@github-manager/utils/locale';
 
-import { NotificationAction } from '../NotificationAdapter';
+import { NotificationAction } from './NotificationAction';
 
 export class DismissAction implements NotificationAction {
     private static readonly LOGGER: Logger = new Logger('notifications::actions::DismissAction');
