+++ src/main/typescript/notifications/actions/NotificationAction.ts
@@ -1,0 +1,7 @@
+import { GitHub } from '@github-manager/client';
+
+export interface NotificationAction {
+    get label(): string;
+
+    execute(notification?: GitHub.Thread): void;
+}
