+++ src/main/typescript/notifications/actions/index.ts
@@ -1,3 +1,4 @@
+export * from './NotificationAction';
 export * from './OpenAction';
 export * from './MarkAsReadAction';
 export * from './DismissAction';
